# docs/api/my-openapi.yml
openapi: 3.1.0
info:
  title: API設計書
  version: 0.0.1
  description: バックエンドエンジニア用研修課題

tags:
  - name: Comment
    description: コメントに関する操作

security: []

servers:
  - url: http://localhost/api

paths:
  # --- コメント登録・コメント一覧 ---
  '/articles/{article_id}/comments':
    parameters:
      - name: article_id
        in: path
        description: 記事のID
        required: true
        schema:
          type: integer
    post: 
      tags:
        - Comment
      summary: コメント登録
      description: 記事にコメントを投稿する
      operationId: storeComment
      requestBody: 
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '200':
          description: 登録成功
        '400':
          description: バリデーションエラー
    get:
      tags:
        - Comment
      summary: コメント一覧
      description: 投稿されたコメントを一覧表示する
      operationId: indexComments
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '404':
          description: 指定された記事が見つかりません
  
  '/articles/{article_id}/comments/{comment_id}':
    parameters:
      - name: article_id
        in: path
        description: 記事のID
        required: true
        schema:
          type: integer

      - name: comment_id
        in: path
        description: コメントのID
        required: true
        schema:
          type: integer
    
    delete:
      tags:
        - Comment
      summary: コメント削除
      description: 投稿されたコメントの削除
      operationId: deleteComments
      responses:
        '200':
          description: 削除成功
        '400':
          description: エラー
    put:
      tags:
        - Comment
      summary: コメント更新
      description: 投稿されたコメントの更新
      operationId: updateComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '200':
          description: 更新成功
        '404':
          description: コメントが見つかりません

components: 
  schemas: 
    # 共通のデータ構造：入力用
    CommentInput:
      type: object
      properties: 
        content:
          type: string
          description: コメントの本文
      required: 
        -  content
    
    # 共通のデータ構造：表示用
    Comment:
      type: object
      properties:
        article_id:
          type: integer
          example: 1
        body:
          type: string
          description: コメント本文
          minLength: 10
          maxLength: 100
          example: "この記事のリスク評価は非常に参考になります。"
        comment_id:
          type: integer
          example: 5
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time